Add 'nonbreaking' to TinyMCE and comment out image resizing
This commit modifies the tinymce.js file. The 'nonbreaking' plugin has been added to the list of plugins and the toolbar. Also, the option to force nonbreaking spaces with the tab key has been enabled. These changes enhance the user's text formatting capabilities. The image resize function has been commented out as it was interfering with the 'autoresize' plugin. It will be revisited later for possible enhancements.